4.7.4 fix(scripts): Test, automation, bisect
- Update test and automation files to align with legacy mapjs repo:
  - Add `test/devtools-recordings/mapjs-node-click.json` to test left
click to select node.
  - `mapjs/scripts/mapjs.env`: Add variable for above script path, plus
one for patch file, applied by bisect script as hotfix, instead of stash
  - `test/test_scripts/mapjs_bisect_testcafe.sh`:
    - Make more reliable.
    - Add flags to simplify testing.
    - Apply patch rather than stash.
    - Add more messages for feedback.
  - `scripts/bash_aliases_mapjs.sh`: Fix so legacy mapjs can use aliases
  - `mapjs/.vscode/settings.json`: Add terminal init file:
`scripts/bash_aliases_mapjs.sh`.

+++ test/test_scripts/mapjs_bisect_testcafe.sh
@@ -3,44 +3,58 @@
 # Script to run for each git bisect iteration (or use alias bru or bre):
 # git bisect run ~/scripts/argmap_test_scripts/mapjs_bisect_testcafe.sh
 
-# First time only (or use alias wss):
+# Otherwise, first time only (or use alias wss):
 # npm exec webpack-dev-server
 
 # Bisect Setup
 # ------------
 
-# QUESTION can I move these two commands outside of this script into calling function?
-# export PATH_REPLAY_SCRIPT=$DIR_PROJECTS/argmap/test/devtools-recordings/add_idea.json
+# shellcheck source=/home/s6mike/scripts/argmap_scripts/bash_aliases_mapjs.sh
+source "$HOME/scripts/argmap_scripts/bash_aliases_mapjs.sh"
 
-# source /home/s6mike/scripts/bash_aliases.sh                      # Functions to ensure clean build.
-# source /home/s6mike/scripts/argmap_scripts/bash_aliases_mapjs.sh # Function to test rendering.
-
-TEST_MODE=false # false allows git commands to remove any changes.
+VALIDATE_PATCH=false # false skips the check that patch needs to work
+TEST_MODE=false      # false allows git commands to remove any changes.
+KEEP_PATCHES=false   # git reset will only happen if this and TEST_MODE is false
+PATH_BISECT_PATCH_FILE=${PATH_BISECT_PATCH_FILE:-"$DIR_PROJECTS/all_mapjs_fixes_3_3_8_modified.diff"}
 
 BISECT_START_TIMESTAMP=${BISECT_START_TIMESTAMP:-$(date "+%Y.%m.%d-%H.%M.%S")}
 
 # Add Hotfixes
 # ------------
 
-# Want stash files to be absolute minimum to avoid merge conflicts
+# Can use stash or diff file, currently using diff.
+# Example diff to generate new patch file:
+#   git diff --no-color e30f8d835e028febe2e951e422c313ac304a0431 HEAD -- . ':(exclude)package-lock.json' ':(exclude).gitignore' ':(exclude)docs/CHANGELOG-mapjs.md' > ../all_mapjs_fixes_3_3_8.diff
+
+# These files are problematic and not being present can break things:
+#   scripts/mapjs.env
+#   .vscode/settings.json
+# Can add them to work area but not staging and git reset --hard won't remove them.
+
 # This stops bisect if apply breaks (125 to skip commit and continue testing)
 #  Alternative to stash apply: https://workflowy.com/#/c1fcaad78669
 
-# Current latest stash fix: All mapjs fixes up to toolbar
-#   /home/s6mike/git_projects/all_mapjs_fixes_to_toolbar.diff
-# Does package.json use 'webpack-dev-server': '^2.4.1'?
-# Has start.js alert but changed to console error? Not essential
-# stash@{0}: On (no branch): package.json version, start.js alert
-# map-model.js _ fix
-# toolbar fix
+# Ensures bisect log is updated each iteration.
+git bisect log >"$PATH_BISECT_LOG"
 
 if [ "$TEST_MODE" = false ]; then # Only runs if not in test mode
 
-  if git stash apply; then
-    echo "'git stash apply' successful."
+  # if git stash apply; then
+  # -v to get verbose feedback
+
+  # Reject allows each patch to be applied independently of success/fails of rest
+  # if git apply -v --ignore-whitespace --3way "$PATH_BISECT_PATCH_FILE"; then
+  if git apply -v --ignore-whitespace --reject "$PATH_BISECT_PATCH_FILE"; then
+    echo "'git apply patch' successful."
   else
-    echo "'git stash apply' failed, exit 255 (abort bisect)"
-    exit 255
+
+    if [ "$VALIDATE_PATCH" = true ]; then
+      echo "'git apply patch' failed, exit 255 (abort bisect)"
+      exit 255
+    else
+      echo "'git apply patch' failed, continuing anyway."
+    fi
+
   fi
 
 fi
@@ -48,21 +62,20 @@ fi
 # Build
 #------
 
+# Should now be available to source
+# shellcheck source=/home/s6mike/git_projects/mapjs/scripts/mapjs.env # Stops shellcheck lint error
+source "$WORKSPACE/scripts/mapjs.env"
+
 # webpack_build
 if [ "$TEST_MODE" = false ]; then # Only runs if not in test mode
   npm install --prefix "$PATH_MJS_HOME" --legacy-peer-deps
 fi
 
-# webpack_server_start # Dev server should already be started
 # git add -u
 
 # TODO: Ideally want to read node_count_before and pass it into the JSON script as a variable to compare it to.
 #   See https://workflowy.com/#/33d0bdfaf875
 # node_count_before = document.getElementsByClassName('mapjs-node').length
-
-# Included in test (expecting 14):
-# node_count_after = document.getElementsByClassName('mapjs-node').length
-# exit if node_count_after > node_count_before then 0 else 1
 
 # Run Test
 # --------
@@ -71,60 +84,52 @@ fi
 
 # Check url:
 # check_url '' || exit_status=125
-
 # if [ $exit_status == 125 ]; then
 #   echo "Page unavailable, setting exit exit_status to 125"
-# else
 
 # Pre-test: Do any elements render?
+# shellcheck disable=SC2119 # No arguments needed, all defaults
 __test_mapjs_renders
 rendered=$?
 
-if [ $rendered = 0 ]; then # headless chrome fails to render any map nodes
-  # Run main test
+echo "rendered: $rendered"
+case $rendered in
+0)
   echo "Render test succeeded, running testcafe script: $PATH_REPLAY_SCRIPT"
   if [ "$PATH_REPLAY_SCRIPT" == "" ]; then
     echo "No TestCafe PATH_REPLAY_SCRIPT set, setting exit status to 255 (bisect abort)"
     exit_status=255
     exit 255 # exits without git reset
   else
-    npm exec testcafe 'chrome:headless --no-default-browser-check' "$PATH_REPLAY_SCRIPT" >>"../git-bisect-logs/$BISECT_START_TIMESTAMP-testcafe-output.txt"
+    npm exec testcafe 'chrome:headless --no-default-browser-check' "$PATH_REPLAY_SCRIPT" >>"$DIR_PROJECTS/git-bisect-logs/$BISECT_START_TIMESTAMP-testcafe-output.txt"
     exit_status=$?
   fi
-else # render succeeds
+  ;;
+
+127)
+  exit_status=255
+  echo "__test_mapjs_renders failure 127: command not found, setting exit status to 255 (bisect abort)"
+  ;;
+
+*)
   exit_status=125
-  echo "Rendering test failed, setting exit status to 125 (bisect skip)."
-fi
+  echo "Rendering test failed, setting exit status to 125 (bisect skip). Is server running? Try 'wss.'"
+
+  # Sanity check rendering in browser
+  # shellcheck disable=SC2119 # No arguments needed, all defaults
+  open-server
+  ;;
+
+esac
 
 # Bisect Teardown
 # ---------------
 
-# node_stop
-# This is necessary to allow checkout to work for next bisect step:
-
 # Only resets if not in test mode
-if [ "$TEST_MODE" = false ]; then # Won't in test mode
+if [ "$TEST_MODE" = false ] && [ "$KEEP_PATCHES" = false ]; then # Won't in test mode
   git reset --hard
 fi
 
 # Return test result
 echo "bisect exit status: $exit_status"
 exit "$exit_status"
-
-# After fix:
-# ----------
-
-# Review bisect log
-# git bisect log >../git-bisect-logs/git_bisect_log_"$(basename --suffix=".json" "$PATH_REPLAY_SCRIPT")".txt
-
-# Repeat bisect
-# git bisect log >../git-bisect-logs/git_bisect_log_"$(basename --suffix=".json" "$PATH_REPLAY_SCRIPT")".txt
-# Edit log, then:
-# git bisect reset
-# git bisect replay ../git-bisect-logs/git_bisect_log_"$(basename --suffix=".json" "$PATH_REPLAY_SCRIPT")".txt
-# git bisect run ~/scripts/argmap_test_scripts/mapjs_bisect_testcafe.sh
-
-# Get diff of staging area vs prev commit
-# git add -u
-# git stash -u -m "cumulative-fixes"
-# git diff --cached >../mapjs_cumulative_fixes.diff
